consolidate how to get supported datatypes
