+++ pandera/api/pandas/types.py
@@ -1,12 +1,15 @@
+# pylint: disable=unused-import
 """Utility functions for pandas validation."""
 
 from functools import lru_cache
-from typing import NamedTuple, Tuple, Type, Union
+from typing import Any, NamedTuple, Type, Union
 
 import numpy as np
 import pandas as pd
 
 from pandera.dtypes import DataType
+from pandera.errors import BackendNotFoundError
+
 
 PandasDtypeInputTypes = Union[
     str,
@@ -17,58 +20,117 @@ PandasDtypeInputTypes = Union[
     np.dtype,
 ]
 
-SupportedTypes = NamedTuple(
-    "SupportedTypes",
-    (
-        ("table_types", Tuple[type, ...]),
-        ("field_types", Tuple[type, ...]),
-        ("index_types", Tuple[type, ...]),
-        ("multiindex_types", Tuple[type, ...]),
-    ),
+PANDAS_LIKE_CLS_NAMES = frozenset(
+    [
+        "DataFrame",
+        "Series",
+        "Index",
+        "MultiIndex",
+        "GeoDataFrame",
+        "GeoSeries",
+    ]
 )
 
 
-@lru_cache(maxsize=None)
-def supported_types() -> SupportedTypes:
-    """Get the types supported by pandera schemas."""
-    # pylint: disable=import-outside-toplevel
-    table_types = [pd.DataFrame]
-    field_types = [pd.Series]
-    index_types = [pd.Index]
-    multiindex_types = [pd.MultiIndex]
+class BackendTypes(NamedTuple):
 
-    try:
-        import pyspark.pandas as ps
+    # list of datatypes available
+    dataframe_datatypes: tuple
+    series_datatypes: tuple
+    index_datatypes: tuple
+    multiindex_datatypes: tuple
+    check_backend_types: tuple
 
-        table_types.append(ps.DataFrame)
-        field_types.append(ps.Series)
-        index_types.append(ps.Index)
-        multiindex_types.append(ps.MultiIndex)
-    except ImportError:
-        pass
-    try:  # pragma: no cover
-        import modin.pandas as mpd
 
-        table_types.append(mpd.DataFrame)
-        field_types.append(mpd.Series)
-        index_types.append(mpd.Index)
-        multiindex_types.append(mpd.MultiIndex)
-    except ImportError:
-        pass
-    try:
+def _get_fullname(obj: Any) -> str:
+    _type = type(obj)
+    return f"{_type.__module__}.{_type.__name__}"
+
+
+@lru_cache
+def get_backend_types(check_cls_fqn: str):
+
+    dataframe_datatypes = []
+    series_datatypes = []
+    index_datatypes = []
+    multiindex_datatypes = []
+
+    mod_name, *mod_path, cls_name = check_cls_fqn.split(".")
+    if mod_name != "pandera":
+        if cls_name not in PANDAS_LIKE_CLS_NAMES:
+            raise BackendNotFoundError(
+                f"cls_name {cls_name} not in {PANDAS_LIKE_CLS_NAMES}"
+            )
+
+    if mod_name == "pandera":
+        # assume mod_path e.g. ["typing", "pandas"]
+        assert mod_path[0] == "typing"
+        *_, mod_name = mod_path
+
+    def register_pandas_backend():
+        from pandera.accessors import pandas_accessor
+
+        dataframe_datatypes.append(pd.DataFrame)
+        series_datatypes.append(pd.Series)
+        index_datatypes.append(pd.Index)
+        multiindex_datatypes.append(pd.MultiIndex)
+
+    def register_dask_backend():
         import dask.dataframe as dd
+        from pandera.accessors import dask_accessor
 
-        table_types.append(dd.DataFrame)
-        field_types.append(dd.Series)
-        index_types.append(dd.Index)
-    except ImportError:
-        pass
+        dataframe_datatypes.append(dd.DataFrame)
+        series_datatypes.append(dd.Series)
+        index_datatypes.append(dd.Index)
 
-    return SupportedTypes(
-        tuple(table_types),
-        tuple(field_types),
-        tuple(index_types),
-        tuple(multiindex_types),
+    def register_modin_backend():
+        import modin.pandas as mpd
+        from pandera.accessors import modin_accessor
+
+        dataframe_datatypes.append(mpd.DataFrame)
+        series_datatypes.append(mpd.Series)
+        index_datatypes.append(mpd.Index)
+        multiindex_datatypes.append(mpd.MultiIndex)
+
+    def register_pyspark_backend():
+        import pyspark.pandas as ps
+        from pandera.accessors import pyspark_accessor
+
+        dataframe_datatypes.append(ps.DataFrame)
+        series_datatypes.append(ps.Series)
+        index_datatypes.append(ps.Index)
+        multiindex_datatypes.append(ps.MultiIndex)
+
+    def register_geopandas_backend():
+        import geopandas as gpd
+
+        register_pandas_backend()
+        dataframe_datatypes.append(gpd.GeoDataFrame)
+        series_datatypes.append(gpd.GeoSeries)
+
+    register_fn = {
+        "pandas": register_pandas_backend,
+        "dask_expr": register_dask_backend,
+        "modin": register_modin_backend,
+        "pyspark": register_pyspark_backend,
+        "geopandas": register_geopandas_backend,
+        "pandera": lambda: None,
+    }[mod_name]
+
+    register_fn()
+
+    check_backend_types = [
+        *dataframe_datatypes,
+        *series_datatypes,
+        *index_datatypes,
+    ]
+
+    return BackendTypes(
+        dataframe_datatypes=tuple(dataframe_datatypes),
+        series_datatypes=tuple(series_datatypes),
+        index_datatypes=tuple(index_datatypes),
+        multiindex_datatypes=tuple(multiindex_datatypes),
+        check_backend_types=tuple(check_backend_types),
     )
 
 
@@ -78,7 +140,9 @@ def is_table(obj):
     Where a table is a 2-dimensional data matrix of rows and columns, which
     can be indexed in multiple different ways.
     """
-    return isinstance(obj, supported_types().table_types)
+    return isinstance(
+        obj, get_backend_types(_get_fullname(obj)).dataframe_datatypes
+    )
 
 
 def is_field(obj):
@@ -87,25 +151,28 @@ def is_field(obj):
     Where a field is a columnar representation of data in a table-like
     data structure.
     """
-    return isinstance(obj, supported_types().field_types)
+    return isinstance(
+        obj, get_backend_types(_get_fullname(obj)).series_datatypes
+    )
 
 
 def is_index(obj):
     """Verifies whether an object is a table index."""
-    return isinstance(obj, supported_types().index_types)
+    return isinstance(
+        obj, get_backend_types(_get_fullname(obj)).index_datatypes
+    )
 
 
 def is_multiindex(obj):
     """Verifies whether an object is a multi-level table index."""
-    return isinstance(obj, supported_types().multiindex_types)
+    return isinstance(
+        obj, get_backend_types(_get_fullname(obj)).multiindex_datatypes
+    )
 
 
 def is_table_or_field(obj):
     """Verifies whether an object is table- or field-like."""
     return is_table(obj) or is_field(obj)
-
-
-is_supported_check_obj = is_table_or_field
 
 
 def is_bool(x):
